move fixture from test/test_step_detect.y file
